requirements.txt
absl-py
inductiva
numpy
pyvista
trimesh
vtk
